fix: 没判断存不存在就检查为不为空

+++ src/models/Pool.ts
@@ -23,6 +23,7 @@ export default class Pool {
     }
     if (!type) return null;
     const fileAbsPath = path.join(this.dir, `${id}.${type.toLowerCase()}`);
+    if (!fs.existsSync(fileAbsPath)) return null;
     if (!fs.statSync(fileAbsPath).size) {
       console.log('删除空文件', fileAbsPath);
       fs.unlinkSync(fileAbsPath);
